+++ CHANGELOG.md
@@ -3,9 +3,12 @@ Written with guidance from [Keep a CHANGELOG](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## v1.2.0
 ### Added
 - parse will take a list of strings
 - JSONPointer.extract returns a list of paths that make up an object 
+- JSONPointer.merge combines dst into src
 
 ## v1.1.0
 ## Added